+++ backend/src/main.ts
@@ -1,5 +1,6 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
+import { ValidationPipe } from '@nestjs/common';
 
 async function bootstrap() {
     const PORT = process.env.PORT || 8888;
@@ -8,5 +9,7 @@ async function bootstrap() {
     await app.listen(PORT, () =>
         console.log(`Server started on port = ${PORT}`),
     );
+
+    app.useGlobalPipes(new ValidationPipe())
 }
 bootstrap();